move AgentStatusChangedEvent to separate package
